chore: :truck: move tests to separate directory
