+++ packages/platforms/react-native/lib/clock.ts
