+++ packages/platforms/react-native/lib/config.ts
