+++ packages/platforms/react-native/lib/index.ts
@@ -1,1 +1,1 @@
-export * from './resource-attributes-source/resource-attributes-source'
+export * from './resource-attributes-source'
