+++ packages/platforms/react-native/tests/clock.test.ts
@@ -1,4 +1,4 @@
-import createClock from './clock'
+import createClock from '../lib/clock'
 
 // reset timer between tests
 beforeEach(() => {
