+++ packages/platforms/react-native/tests/config.test.ts
@@ -1,4 +1,4 @@
-import { createSchema } from './config'
+import { createSchema } from '../lib/config'
 
 const stringValidation = [
   { expected: false, value: true },
